fix(workflow): Set default for WorkflowState.scope_data to None
Changed the default value for the `scope_data` JSONField from `dict`
to `None`.

The previous default of `dict` causes issues with mutable default
arguments when creating new instances in Django. Setting it to `None`
ensures a proper `null=True` behavior, which is safer for JSONFields
that are not immediately populated.